% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{add_col,Dataset-method}
\alias{add_col,Dataset-method}
\title{Añadir una columna}
\usage{
\S4method{add_col}{Dataset}(object, data, name = NULL)
}
\arguments{
\item{object}{El Dataset que se quiere modificar}

\item{data}{Vector de valores de la columna a añadir. Todos los valores tienen que ser del mismo tipo.}

\item{name}{Nombre de la columna, opcional}
}
\value{
Un nuevo objeto de clase Dataset con la columna añadida
}
\description{
Método para añadir una columna al Dataset
}
\examples{
dataset = new("Dataset")
dataset = add_col(dataset,c(1,2,3))
dataset = add_col(dataset,c("Jon","Bill","James"))
dataset = add_col(dataset,c(7,8,9))
dataset

}
