% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{add_cols,Dataset,list-method}
\alias{add_cols,Dataset,list-method}
\title{Añadir varias columnas}
\usage{
\S4method{add_cols}{Dataset,list}(object, data, names = NULL)
}
\arguments{
\item{object}{El Dataset que se quiere modificar}

\item{data}{Lista de vectores que contienen los valores de las columnas a añadir}

\item{names}{Vector de nombres de las columnas, opcional}
}
\value{
Un nuevo objeto de clase Dataset con la columnas añadida
}
\description{
Método para añadir varias columnas al Dataset
}
\examples{
dataset = new("Dataset")
dataset = add_cols(dataset,list(c(1,2,3),c("Jon","Bill","James"),c(7,8,9))
dataset

}
