% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{add_row,Dataset,list-method}
\alias{add_row,Dataset,list-method}
\title{Añadir una fila}
\usage{
\S4method{add_row}{Dataset,list}(object, data)
}
\arguments{
\item{object}{El Dataset que se quiere modificar}

\item{data}{Vector de valores de la fila a añadir.}
}
\value{
Un nuevo objeto de clase Dataset con la fila añadida
}
\description{
Método para añadir una fila al Dataset
}
\examples{
dataset = new("Dataset")
dataset = add_row(dataset,c(1,"Bill",5))
dataset = add_row(dataset,c(2,"Jon",6))
dataset

}
