% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{add_rows,Dataset,list-method}
\alias{add_rows,Dataset,list-method}
\title{Añadir varias filas}
\usage{
\S4method{add_rows}{Dataset,list}(object, data)
}
\arguments{
\item{object}{El Dataset que se quiere modificar}

\item{data}{Lista de listas que contienen los valores de las filas a añadir}
}
\value{
Un nuevo objeto de clase Dataset con las filas añadidas
}
\description{
Método para añadir varias filas al Dataset
}
\examples{
dataset = new("Dataset")
dataset = add_rows(dataset,list(c(1,"Bill",5),c(2,"Jon",6)))
dataset

}
