% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{AUC}
\alias{AUC}
\title{Valor AUC}
\usage{
AUC(coords.list)
}
\arguments{
\item{coords.list}{Lista de 2 elementos: coordenadas X e Y de los puntos que conforman la curva ROC}
}
\value{
Valor que corresponde al AUC
}
\description{
Esta función calcula el área de debajo de la curva ROC, es decir,
el valor AUC (Area Under the Curve)
}
\examples{
X <- c(3.5,7.4,1.2,2.0,3.9,5.3,6.1,6.9)
Y <- c(TRUE,TRUE,FALSE,FALSE,FALSE,FALSE,TRUE,TRUE)
df <- data.frame(X,Y)
coords.list = ROC(df,"X","Y")
AUC(coords.list)
}
