% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{boxplot_dataset}
\alias{boxplot_dataset}
\title{Boxplot}
\usage{
boxplot_dataset(x, together = TRUE)
}
\arguments{
\item{x}{Dataframe a visualizar}

\item{together}{Valor boolear que especifica si se quieren ver todas las variables en un mismo boxplot o no}
}
\description{
Esta función visualiza los valores numéricos de un dataframe en un gráfico de tipo boxplot.
}
\examples{
Name <- c("Jon", "Bill", "Maria", "Ben", "Tina")
Age <- c(23, 41, 32, 58, 26)
Weight <- c(156,98,75,112,74)
Height <- c(1.75,1.68,1.56,1.86,1.76)
Hair = c("Long", "Short","Long", "Short","Long")
Independent = c(TRUE, FALSE, TRUE, TRUE, FALSE)
x <- data.frame(Name, Age, Weight, Height, Hair, Independent)
boxplot_dataset(x,together=FALSE)
}
