% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{conditional.entropy}
\alias{conditional.entropy}
\title{Entropía condicional entre vectores}
\usage{
conditional.entropy(X, Y)
}
\arguments{
\item{X}{Primer vector (condicionado)}

\item{Y}{Segundo vector (condicionante)}
}
\value{
Valor numérico de la entropía condicional
}
\description{
Esta función calcula la entropía condicional entre 2 vectores
}
\examples{
Name <- c("Jon", "Bill", "Maria", "Ben", "Tina")
Age <- c(23, 41, 32, 58, 26)
conditional.entropy(Name,Age)
}
