% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlmutua.R
\name{correlacion}
\alias{correlacion}
\title{Correlación de Pearson entre columnas}
\usage{
correlacion(x)
}
\arguments{
\item{x}{Matriz o dataframe sobre el cual calcular las correlaciones}
}
\value{
Vector numérico de las correlaciones entre las columnas numéricas de \code{x}
}
\description{
Esta función calcula los coeficientes de correlación de Pearson
de entre todos los pares de columnas numéricas de un array
}
\examples{
Name <- c("Jon", "Bill", "Maria", "Ben", "Tina")
Age <- c(23, 41, 32, 58, 26)
Weight <- c(156,98,75,112,74)
Height <- c(1.75,1.68,1.56,1.86,1.76)
x <- data.frame(Name, Age, Weight, Height)
correlacion(x)
}
