% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwritedf.R
\name{dataframe2csv}
\alias{dataframe2csv}
\title{Guardado de dataframe en fichero CSV}
\usage{
dataframe2csv(x, file.path)
}
\arguments{
\item{x}{Dataframe a guardar.}

\item{file.path}{Dirección y nombre del fichero de tipo CSV en donde guardar los datos}
}
\description{
Esta función escribe y guarda los datos de un dataframe de entrada en el
fichero tipo CSV especificado. El formato es el siguiente:

 ,nombrecolumna1,nombrecolumna2,nombrecolumna3...
 nombrefila1,datos,de,fila,1,separados,por,coma
 nombrefila2,datos,de,fila,2,separados,por,coma
 nombrefila3,datos,de,fila,3,separados,por,coma
 ...
}
\examples{
data = data.frame(
Nombre = c("Alice", "Bob", "Charlie"),
Edad = c(25, 30, 35),
Ciudad = c("Nueva York", "Los Ángeles", "Chicago")
)
rownames(data) = c("Primero","Segundo","Tercero")
dataframe2csv(data, "ejemplo.csv")
}
