% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset_from_table}
\alias{dataset_from_table}
\title{Constructor avanzado de la clase Dataset}
\usage{
dataset_from_table(x)
}
\arguments{
\item{x}{Tabla de datos (matriz o data frame)}
}
\value{
Un objeto de clase \code{\linkS4class{Dataset}} que representa la tabla de datos pasada como argumento
}
\description{
Esta función crea un objeto de clase \code{\linkS4class{Dataset}} a partir de una tabla de datos (matriz o data.frame)
}
