% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{discretize}
\alias{discretize}
\title{Discretizar con puntos de corte}
\usage{
discretize(x, cut.points)
}
\arguments{
\item{x}{Vector numérico a discretizar}

\item{cut.points}{Vector de puntos de corte}
}
\value{
Vector tipo factor con los valores de \code{x} discretizados en intérvalos
}
\description{
Esta función discretiza los valores de un vector dados los puntos de corte
}
\examples{
x <- c(3.5,7.4,1.2,2.0,3.9,5.3,6.1,6.9)
discretize(x, c(2,4,5.5))
}
