% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{discretizeEF_vector}
\alias{discretizeEF_vector}
\title{Discretización de "equal frequency" para vectores}
\usage{
discretizeEF_vector(x, num.bins, return.cut.points = FALSE)
}
\arguments{
\item{x}{Vector numérico sobre el que aplicar el algoritmo}

\item{num.bins}{Número de intervalos con los que discretizar}

\item{return.cut.points}{Si se deben devolver los puntos de corte (\code{TRUE}) o no (\code{FALSE} por default)}
}
\value{
Si \code{return.cut.points==TRUE} devuelve una lista con 2 elementos:
 1) Vector tipo factor del mismo tamaño que \code{x}, con los valores numéricos discretizados en intérvalos
 2) Puntos de corte de los intérvalos
 Si \code{return.cut.points==FALSE}, devolverá un vector tipo factor
}
\description{
Esta función discretiza todos los valores numéricos
de un vector siguiendo el algoritmo "equal frequency"
}
\examples{
x <- rnorm(8,3,5)
discretizeEF_vector(x,5)
}
