% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{discretizeEW}
\alias{discretizeEW}
\title{Discretización de "equal width" para arrays}
\usage{
discretizeEW(x, num.bins)
}
\arguments{
\item{x}{Vector numérico, matriz numérica o un dataframe sobre el que aplicar el algoritmo}

\item{num.bins}{Número de intervalos con los que discretizar}
}
\value{
Array del mismo tamaño que \code{x}, con los valores numéricos discretizados en intérvalos
}
\description{
Esta función discretiza todos los valores numéricos
de un vector o una tabla siguiendo el algoritmo "equal width"
}
\examples{
x <- matrix(rnorm(30,5,10),ncol=3)
discretizeEW(x,5)
}
