% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{entropy.binning}
\alias{entropy.binning}
\title{Discretización por entropía para vectores}
\usage{
entropy.binning(X, Y, num.bins, tries = 5, random.cut = FALSE, offset = NA)
}
\arguments{
\item{X}{Vector numérico a discretizar}

\item{Y}{Vector respecto al cual se quiere maximizar la información mutua}

\item{num.bins}{Número de intervalos con los que discretizar}

\item{tries}{Cuantos intentos de discretización se harán con distintos puntos de corte}

\item{random.cut}{Valor boolear que especifica si los puntos de corte se calcularán de manera aleatoria o no}

\item{offset}{En el caso de que no haya aleatoriedad, en cada intento se incrementarán los puntos de corte por este valor numérico. Por defecto es None, y el offset será la desviación estándar de los datos dividido por el numéro de intentos}
}
\value{
Lista con 3 valores:
    1- pandas.core.arrays.categorical.Categorical con los valores de x discretizados en intérvalos
    2- Lista de puntos de corte calculados
    3- El valor de información mutua máxima obtenida
}
\description{
Esta función discretiza los valores de un vector con el objetivo de maximizar la información mutua
respecto a otra variable.
}
\examples{
X = c(23, 41, 32, 58, 26)
Y = c(TRUE, FALSE, TRUE, TRUE, FALSE)
entropy.binning(X,Y,num.bins=3,tries=10,random.cut=TRUE)
}
