% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{filtrar}
\alias{filtrar}
\title{Filtrado de variables}
\usage{
filtrar(x, condition, use.entropy = TRUE, use.varianza = FALSE)
}
\arguments{
\item{x}{Matriz o dataframe que se quiere filtrar}

\item{condition}{Condición que debe cumplir la métrica de una columna
para que se conserve}

\item{use.entropy}{Si se quiere utilizar la entropía como métrica de filtrado
(\code{TRUE} por defecto)}

\item{use.varianza}{Si se quiere utilizar la varianza como métrica de filtrado
(\code{FALSE} por defecto)}
}
\value{
Matriz o dataframe con las columnas filtradas
}
\description{
Esta función elimina las columnas de una matriz o dataframe
que no cumplan una condición dada respectiva a su varianza o entropía.
}
\examples{
Name <- c("Jon", "Bill", "Maria", "Ben", "Tina")
Age <- c(23, 41, 32, 58, 26)
Weight <- c(156,98,75,112,74)
Height <- c(1.75,1.68,1.56,1.86,1.76)
Working <- c(FALSE, TRUE, FALSE, TRUE, TRUE)
x <- data.frame(Name, Age, Weight, Height, Working)
filtrar(x, condition="< 2",use.varianza = TRUE, use.entropy = TRUE)
}
