% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_length,Dataset-method}
\alias{get_length,Dataset-method}
\title{Tamaño del Dataset}
\usage{
\S4method{get_length}{Dataset}(object)
}
\arguments{
\item{object}{El Dataset que se quiere medir}
}
\value{
Un integer que indica el número de columnas del Dataset
}
\description{
Método para obtener el tamaño del Dataset
}
\examples{
dataset = new("Dataset")
dataset = add_col(dataset,c(1,2,3))
dataset = add_col(dataset,c("Jon","Bill","James"))
get_length(dataset)
}
