% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{getitem,Dataset-method}
\alias{getitem,Dataset-method}
\title{Indexar un Dataset}
\usage{
\S4method{getitem}{Dataset}(object, i, j = NULL)
}
\arguments{
\item{object}{El Dataset que se quiere indexar}

\item{i}{índice o nombre de la columna a indexar}

\item{j}{índice de la fila a indexar, opcional}
}
\value{
Un vector con los valores de la columna o un valor concreto si se especifica el segundo índice
}
\description{
Método para indexar un Dataset. El primer índice indica la columna y el segundo la fila.
Si no se especifica el segundo índice, se devuelve la columna completa.
}
\examples{
dataset = new("Dataset")
dataset = add_col(dataset,c(1,2,3))
dataset = add_col(dataset,c("Jon","Bill","James"))
getitem(dataset,"V1")
getitem(dataset,"V1",3)
getitem(dataset,1,1)

}
