% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{is.binary}
\alias{is.binary}
\title{Revisar si un vector es binario}
\usage{
is.binary(x)
}
\arguments{
\item{x}{Vector que revisar}
}
\value{
\code{TRUE} si \code{x} es binario, \code{FALSE} en el caso contrario
}
\description{
Esta función revisa si un vector de entrada es binario o no
}
\examples{
X <- c(TRUE,TRUE,FALSE,FALSE,FALSE,FALSE,TRUE,TRUE)
is.binary(X)
X <- c(3.5,7.4,1.2,2.0,3.9,5.3,6.1,6.9)
is.binary(X)
}
