% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metricas}
\alias{metricas}
\title{Métricas de tablas}
\usage{
metricas(x, attribute.variable = NA, predict.variable = NA)
}
\arguments{
\item{x}{Matriz numérica o dataframe que contiene los datos}

\item{attribute.variable}{Índice o nombre de la columna que corresponde a los atributos
que se utilizan para hacer la predicción (\code{NA} por defecto)}

\item{predict.variable}{Índice o nombre de la columna binaria que se quiere predecir
(\code{NA} por defecto)}
}
\value{
Vector numérico que contiene todas las métricas
}
\description{
Esta función calcula las varianzas de las columnas numéricas y
las entropías de las categóricas de una matriz o dataframe. Si se especifican
una variable predictora y una columna objetivo, también calcula la curva ROC
y el valor AUC.
}
\examples{
Name <- c("Jon", "Bill", "Maria", "Ben", "Tina")
Age <- c(23, 41, 32, 58, 26)
Weight <- c(156,98,75,112,74)
Height <- c(1.75,1.68,1.56,1.86,1.76)
Working <- c(FALSE, TRUE, FALSE, TRUE, TRUE)
x <- data.frame(Name, Age, Weight, Height, Working)
metricas(x,"Age","Working")
}
