% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlmutua.R
\name{mutual.correlation}
\alias{mutual.correlation}
\title{Matriz de correlación e información mutua}
\usage{
mutual.correlation(x)
}
\arguments{
\item{x}{Matriz o dataframe sobre el cual calcular la información mutua y las correlaciones}
}
\value{
Matriz numérica de los valores de información mutua y correlación entre las columnas de \code{x}
}
\description{
Esta función calcula la información mutua entre las columnas
categóricas de una matriz o dataframe, y la correlación entre las numéricas
}
\examples{
Name <- c("Jon", "Bill", "Maria", "Ben", "Tina")
Age <- c(23, 41, 32, 58, 26)
Weight <- c(156,98,75,112,74)
Height <- c(1.75,1.68,1.56,1.86,1.76)
x <- data.frame(Name, Age, Weight, Height)
mutual.correlation(x)
}
