% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlmutua.R
\name{mutual.information}
\alias{mutual.information}
\title{Información mutua entre columnas}
\usage{
mutual.information(x, only.categorical = FALSE)
}
\arguments{
\item{x}{Matriz o dataframe sobre el cual calcular la información mutua}

\item{only.categorical}{Si calcular la información mutua entre todas las columnas (\code{FALSE} por default) o sólo con las categóricas (\code{TRUE})}
}
\value{
Vector numérico de los valores de información mutua entre las columnas de \code{x}
}
\description{
Esta función calcula la información mutua entre las columnas
de una matriz o dataframe
}
\examples{
Name <- c("Jon", "Bill", "Maria", "Ben", "Tina")
Age <- c(23, 41, 32, 58, 26)
Weight <- c(156,98,75,112,74)
Height <- c(1.75,1.68,1.56,1.86,1.76)
x <- data.frame(Name, Age, Weight, Height)
mutual.information(x)
}
