% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlmutua.R
\name{mutual.information.vectors}
\alias{mutual.information.vectors}
\title{Información mutua entre 2 vectores}
\usage{
mutual.information.vectors(X, Y)
}
\arguments{
\item{X}{Vector 1 sobre el cual calcular la información mutua}

\item{Y}{Vector 2 sobre el cual calcular la información mutua}
}
\value{
Valor numérico que corresponde a la información mutua entre los 2 vectores de entrada
}
\description{
Esta función calcula la información mutua entre 2 vectores
}
\examples{
Name <- c("Jon", "Bill", "Maria", "Ben", "Tina")
Age <- c(23, 41, 32, 58, 26)
Weight <- c(156,98,75,112,74)
Height <- c(1.75,1.68,1.56,1.86,1.76)
x <- data.frame(Name, Age, Weight, Height)
mutual.information(x)
}
