% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stannorm.R
\name{normalize}
\alias{normalize}
\title{Normalizar arrays}
\usage{
normalize(x)
}
\arguments{
\item{x}{Vector, matriz o dataframe que normalizar}
}
\value{
Array normalizado de la misma dimensión que \code{x}
}
\description{
Esta función normaliza los valores de un vector y las columnas de
una matriz o dataframe para que esten en un rango [0,1]
}
\examples{
x <- matrix(rnorm(30,5,10),ncol=3)
normalize(x)
}
