% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{numeric.columns}
\alias{numeric.columns}
\title{Buscar columnas numéricas}
\usage{
numeric.columns(x)
}
\arguments{
\item{x}{Matriz o dataframe que analizar}
}
\value{
Un vector lógico que especifica que columnas son numéricas (TRUE)
y cuales no (FALSE)
}
\description{
Esta función busca columnas numéricas en una matriz o dataframe
}
\examples{
Name <- c("Jon", "Bill", "Maria", "Ben", "Tina")
Age <- c(23, 41, 32, 58, 26)
Weight <- c(156,98,75,112,74)
Height <- c(1.75,1.68,1.56,1.86,1.76)
x <- data.frame(Name, Age, Weight, Height)
numeric.columns(x)
}
