% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{pieplot}
\alias{pieplot}
\title{Pie plot}
\usage{
pieplot(x, only.categorical = TRUE)
}
\arguments{
\item{x}{Dataframe a visualizar}

\item{only_categorical}{Valor boolear que especifica si sólo se quieren visualizar sólo las variables categóricas o no}
}
\description{
Esta función visualiza las variables de un dataframe distintos gráficos tipo pieplot ("tarta").
}
\examples{
Name <- c("Jon", "Bill", "Maria", "Ben", "Tina")
Age <- c(23, 41, 32, 58, 26)
Weight <- c(156,98,75,112,74)
Height <- c(1.75,1.68,1.56,1.86,1.76)
Hair = c("Long", "Short","Long", "Short","Long")
Independent = c(TRUE, FALSE, TRUE, TRUE, FALSE)
x <- data.frame(Name, Age, Weight, Height, Hair, Independent)
pieplot(x,only.categorical=TRUE)
}
