% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwritedf.R
\name{read.dataframe}
\alias{read.dataframe}
\title{Lectura de datos de un fichero}
\usage{
read.dataframe(file.path)
}
\arguments{
\item{file.path}{Dirección y nombre del fichero a leer}
}
\value{
Matriz o dataframe construido a partir de los datos leídos en el
fichero. Si el dataset está nombrado, devolverá una lista donde el primer
elemento especificará el nombre y el segundo el dataset
}
\description{
Esta función lee un dataset desde un fichero especificado que haya sido
creado con el método gemelo \code{write.dataframe}
}
\examples{
Name = c("Jon", "Bill", "Maria", "Ben", "Tina")
Age = c(23, 41, 32, 58, 26)
Weight = c(156,98,75,112,74)
Height = c(1.75,1.68,1.56,1.86,1.76)
x = data.frame(Name, Age, Weight, Height)
write.dataframe(list("MiDataset",x),"custom_data.txt")
read.dataframe("custom_data.txt")
}
