% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{remove_col,Dataset-method}
\alias{remove_col,Dataset-method}
\title{Quitar una columna}
\usage{
\S4method{remove_col}{Dataset}(object, i)
}
\arguments{
\item{object}{El Dataset que se quiere modificar}

\item{i}{índice o nombre de la columna a quitar}
}
\value{
Un nuevo objeto de clase Dataset con la columna quitada
}
\description{
Método para quitar una columna del Dataset
}
\examples{
dataset = new("Dataset")
dataset = add_col(dataset,c(1,2,3),name="nums")
dataset = add_col(dataset,c("Jon","Bill","James"),name="names")
dataset = remove_col(dataset,1)
dataset
dataset = remove_col(dataset,"names")
dataset
}
