% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{remove_cols,Dataset,list-method}
\alias{remove_cols,Dataset,list-method}
\title{Quitar varias columnas}
\usage{
\S4method{remove_cols}{Dataset,list}(object, l)
}
\arguments{
\item{object}{El Dataset que se quiere modificar}

\item{l}{lista de índices o nombres de las columnas a quitar}
}
\value{
Un nuevo objeto de clase Dataset con las columnas quitadas
}
\description{
Método para quitar varias columnas del Dataset
}
\examples{
dataset = new("Dataset")
dataset = add_col(dataset,c(1,2,3),name="nums")
dataset = add_col(dataset,c("Jon","Bill","James"),name="names")
dataset = remove_cols(dataset,list(1,"names"))
dataset
}
