% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{rename_cols,Dataset,character-method}
\alias{rename_cols,Dataset,character-method}
\title{Renombrar columnas del Dataset}
\usage{
\S4method{rename_cols}{Dataset,character}(object, new_names)
}
\arguments{
\item{object}{El Dataset que se quiere modificar}

\item{new_names}{Vector de carácteres que recoge los nuevos nombres}
}
\value{
Un nuevo objeto de clase Dataset con las columnas renombradas
}
\description{
Método para renombrar las columnas de un Dataset
}
\examples{
dataset = new("Dataset")
dataset = add_col(dataset,c(1,2,3))
dataset = add_col(dataset,c("Jon","Bill","James"))
dataset = rename_cols(dataset,c("Index","Name"))
dataset
}
