% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{ROC}
\alias{ROC}
\title{Curva ROC}
\usage{
ROC(x, attribute.variable, predict.variable)
}
\arguments{
\item{x}{Todos los datos recogidos en una matriz numérica o dataframe}

\item{attribute.variable}{Índice o nombre de la columna que corresponde a los atributos
numéricos que se utilizan para hacer la predicción}

\item{predict.variable}{Índice o nombre de la columna binaria que se quiere predecir}
}
\value{
Lista con 2 elementos:
 1) Coordenadas del eje X de la curva ROC (valores FPR)
 2) Coordenadas del eje Y de la curva ROC (valores TPR)
}
\description{
Esta función calcula los puntos de la curva ROC para
la predicción de una variable binaria y visualiza el gráfico
}
\examples{
X <- c(3.5,7.4,1.2,2.0,3.9,5.3,6.1,6.9)
Y <- c(TRUE,TRUE,FALSE,FALSE,FALSE,FALSE,TRUE,TRUE)
df <- data.frame(X,Y)
ROC(df,1,2)
}
