% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stannorm.R
\name{standarize}
\alias{standarize}
\title{Estandarizar arrays}
\usage{
standarize(x)
}
\arguments{
\item{x}{Vector, matriz o dataframe que estandarizar}
}
\value{
Array estandarizado de la misma dimensi칩n que \code{x}
}
\description{
Esta funci칩n estandariza los valores de un vector y las columnas de
una matriz o dataframe para que la media sea 0 y la desviaci칩n est치ndar 1
}
\examples{
x <- matrix(rnorm(30,5,10),ncol=3)
standarize(x)
}
