% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{to_dataframe,Dataset-method}
\alias{to_dataframe,Dataset-method}
\title{Dataset a data.frame}
\usage{
\S4method{to_dataframe}{Dataset}(object)
}
\arguments{
\item{object}{El Dataset que se quiere convertir}
}
\value{
Un data.frame con los mismos datos que el Dataset de entrada
}
\description{
Método para convertir un Dataset a un data.frame
}
\examples{
dataset = new("Dataset")
dataset = add_col(dataset,c(1,2,3))
dataset = add_col(dataset,c("Jon","Bill","James"))
dataset = add_col(dataset,c(7,8,9))
df = to_dataframe(dataset)
df

}
