% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{varianza}
\alias{varianza}
\title{Varianza de arrays}
\usage{
varianza(x)
}
\arguments{
\item{x}{Array de valores sobre los cuales calcular la varianza}
}
\value{
En el caso de que \code{x} sea un vector, devolverá el valor de varianza.
Si no, devolverá un vector que recoja las varianzas de las columnas numéricas de \code{x}
}
\description{
Esta función calcula la varianza de un vector numérico o de
cada columna numérica de un matriz o dataframe
}
\examples{
x <- matrix(rnorm(30,5,10),ncol=3)
varianza(x)
}
