% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwritedf.R
\name{write.dataframe}
\alias{write.dataframe}
\title{Guardado de datos en fichero}
\usage{
write.dataframe(x, file.path)
}
\arguments{
\item{x}{Matriz o dataframe a guardar. También puede ser una lista, donde
el primer elemento especifica el nombre y el segundo el dataset a guardar}

\item{file.path}{Dirección y nombre del fichero en donde guardar los datos}
}
\description{
Esta función escribe y guarda una matriz o dataframe de entrada en el
fichero especificado. El formato es el siguiente (los símbolos <> indican que
la información es opcional):

tipodeobjeto,númerofilas,númerocolumnas<,nombre>
<#Rownames: nombrefila1,nombrefila2,nombrefila3...>
<#Colnames: nombrecolumna1,nombrecolumna2,nombrecolumna3...>
datos,de,columna,1,separados,por,coma
datos,de,columna,2,separados,por,coma
datos,de,columna,3,separados,por,coma
...
}
\examples{
Name = c("Jon", "Bill", "Maria", "Ben", "Tina")
Age = c(23, 41, 32, 58, 26)
Weight = c(156,98,75,112,74)
Height = c(1.75,1.68,1.56,1.86,1.76)
x = data.frame(Name, Age, Weight, Height)
write.dataframe(list("MiDataset",x),"custom_data.txt")
}
